worker2/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
asyncssh==2.22.0
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
et_xmlfile==2.0.0
frozenlist==1.8.0
grpcio==1.78.0
grpcio-health-checking==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.0
pdf2image==1.17.0
pillow==11.3.0
propcache==0.4.1
protobuf==6.33.5
pycparser==3.0
pymssql==2.3.13
pytesseract==0.3.13
python-dotenv==1.2.2
pyzeebe==4.7.0
requests==2.32.5
requests-oauthlib==2.0.0
typing_extensions==4.15.0
urllib3==2.6.3
xlrd==2.0.2
yarl==1.23.0
